\name{CMulti}
\docType{data}
\alias{CMulti}
\title{Microarray data set of Su on different mammalian tissue types}
\description{
  Microarray data from Broad
  Institute \dQuote{Cancer Program Data Sets} which
  was produced by Su et al. 2002.

  Gene expression from human and mouse samples across a diverse array
  of tissues, organs, and cell lines have been profiled.
  The goal was to have a reference for the normal mammalian transcriptome.

  Here we want to identify the subclasses which correspond to the tissue
  types.


}
\usage{CMulti}
\format{  Vector \code{CMulti} of 102 samples giving the
  four classes of tissue types.}
\source{Broad Institute
  \dQuote{Cancer Program Data Sets}:
  \url{http://www.broadinstitute.org/cgi-bin/cancer/datasets.cgi}}
\references{
Hoshida Y, Brunet J-P, Tamayo P, Golub TR, Mesirov JP,
\sQuote{Subclass Mapping: Identifying Common Subtypes
  in Independent Disease Data Sets},
PLoS ONE 2(11): e1195, 2007.

Su AI, Cooke MP, Ching KA, Hakak Y, Walker JR, et al.
\sQuote{Large-scale analysis of the human and mouse transcriptomes},
Proc. Natl. Acad. Sci. USA 99:4465-4470, 2002.
}
\keyword{datasets}


