\name{fabiaDataVersion}
\alias{fabiaDataVersion}
\title{Display version info for package fabiaData}
\description{
  \code{fabiaDataVersion} displays version information about the
  package.
}
\usage{
fabiaDataVersion()
}
\seealso{
\code{\link{Breast_A}},
\code{\link{DLBCL_B}},
\code{\link{Multi_A}},
}
\author{Sepp Hochreiter}
\examples{
fabiaDataVersion()
}
\keyword{models}

